fix: Fix per-file config interaction with one py_binary per main (#1664)
[This previous PR](https://github.com/bazelbuild/rules_python/pull/1584)
added the ability to make a `py_binary` target per file if `if __name__
== "__main__"` tokens were found in the file. This works great in the
default case, but when `python_generation_mode` is set to `file`, the
plugin now attempts to make both a `py_binary` and a `py_library` target
for each main file, which results in an error.

This PR modifies the behavior to work properly with per-file target
generation, and adds tests for this case.