refactor(pip_install): make the API surface of deps parsing smaller (#1657)
With this change we also fix a minor bug where the self edge
dependencies
within the wheels are correctly handled as per unit tests.

Also refactor the code to use the fact that the selects are chosen
based on the specificity of the condition. This can make the addition
of deps incrementally correct without the need for extra processing in
the `build` method. This should hopefully make the implementation more
future-proof as it makes use of these generic concepts and is likely to
help to extend this code to work with selecting on multiple Python
versions.

I have added a test to ensure coverage of the new code.