fix(bzlmod): make import rules_python.python.runfiles work under bzlmod
In bzlmod, the name of the `@rules_python` repo directory name in the
runfiles tree changed from simply `rules_python` to
`rules_python~VERSION`, where "VERSION" includes the module version
number and some other parts. This broke `import rules_python`. Worse,
the name isn't knowable in advance and isn't a valid Python package
name, so making it importable is very difficult. This makes it harder to
switch to using bzlmod.

As a short term solution, we made `import python` work, which works but
(1) requires users to update their code, and (2) requires us to take the
more generic "python" top-level name, neither of which are things we
want.

To fix, create an importable directory named "rules_python" and set it
up so that `import rules_python.python.runfiles` works.

This importable `rules_python` directory is created under a `src`
sub-directory so that we can better control our public interface, and,
eventually, reduce the number of sys.path entries needed to just that
single `src` directory.

For now, it's kept simple: `rules_python` just imports
`pythons.runfiles` and patches up sys.modules to make it appear as-if it
was an actual sub-module. This is done so that `import python.runfiles`
still works and looks the same (e.g. same `__file__` etc).

This unfortunately means an extra entry on sys.path (`$repoRoot` is
added by `//python/runfiles:runfiles`; `$repoRoot/src` is added by
`//src/rules_python:__init__`). Fixing that can be done in a future
change by moving the files around and doing some different import
patching.