feat: add interpreter_version_info to py_runtime (#1671)
Adds an `interpreter_version_info` attribute to the `py_runtime` and
associated provider that maps to the `sys.version_info` values. This
allows the version of the interpreter to be known statically, which can
be useful for rule sets that depend on the interpreter, and need to
build environments / pathing that contain version info (virtualenvs for
example).

---------

Co-authored-by: Richard Levasseur <rlevasseur@google.com>