chore: remove MODULE.bazel.lock and ignore it (#1680)
When we use the latest `bazel` release in the `rules_python`
module, then the lock file gets created, but I am not sure if
we should commit it in. Since the lock file is very useful
for local development as it speeds up dependency refetching,
this PR disables it as advised in the issue below. Whilst at
it, the `bzlmod` example also removes it for smaller diffs
when we develop extensions.

See bazelbuild/bazel#20369