feat: support resolving deps for a non-host interpreter version
With this PR we prepare the dependency resolution machinery to
support other than host python versions. This may make it possible
to generate select statements based on the version but we target
this usecase as it requires little code to be changed and once
wired to the starlark code could be of use. Creating this PR
for discussion so that we can decide on a suitable interface.