refactor: broaden visibility and use list() instead of keys() (#1704)
This upstreams a couple trivial Google patches to make rules_python more
compatible with some Google-internal changes.

* Use list(dict) instead of dict.keys(). This is because some bzl files
are run through a Python-based testing framework, and dict.keys() can't
be concatenated to a list in Python 3.
* Expand visibility to all of rules_python instead of just the python
subdirectory. This makes some patches that use internals easier to
maintain, but also makes the visiblity more in line with the rest of the
project (where `//:__subpackages__` is used for convenience, as it
avoids having to change visibility frequently, but still prevents public
dependencies).