feat(py_wheel): Added `requires_file` and `extra_requires_files` attrs (#1710)
The `compile_pip_requirements` rule promotes having `requirements.in`
files describe python dependencies. This change aims to allow these
files to be the source of truth for constraints by allowing the
`py_wheel` rule to use them for adding requirements to a wheel. This
reduces overhead in needing to maintain two lists of equal information
(one as he `.in` and the other as starlark data).