fix: make windows ignore pycs in runtime (#1715)
The downloaded runtimes don't contain pyc files, which means Python
tries to create them at runtime, and thus the set of files in the
runtime changes. This normally just silently affects cache hits, but on
Windows can result is errors. This is because Windows has stricter file
access semantics (e.g. a file can't be deleted if its still opened) and
the way Python programs are built involves a step to zip everything up
(which then requires the pyc to exist to be put into the zip file).

To fix, always ignore pyc files for Windows hosts. This should prevent
those files from being picked up by the glob and being considered part
of the runtime.

Co-authored-by: aignas <240938+aignas@users.noreply.github.com>