fix(bzlmod): do not fail pip.parse with empty requirements.txt (#1718)
With this change we handle correctly the case where the requirements.txt
might be empty (e.g. during onboarding to `rules_python`). I have added
an integration test because this usecase seems important enough to be
covered
to ensure that we have no regressions. It also creates a place to add
more
`pip.parse` and `pip_parse` tests.

Closes #1623.