Remove envsubst from _pip_repository_impl
In the case where you vendor the requirements.bzl file, the
environment variables at the time the requirements.bzl file was
generated were being used. It is much more likely that the desired
behavior is to use the environment variables at the time the
dependencies are fetched, apart from the fact that some environment
variables may contain sensitive info.

There's already an envsubst in _whl_library_impl, so we just need to
propagate the envsubst list in the _config we write to the vendored
requirements.bzl file.