repro: create a small repo rule for python interpreter resolution and third party deps
Use this by running `bazel build @repro//..` in the `rules_python` module.
The build will fail all the time because I do `sys.exit(1)` in the
script but the expectation that on Windows we can reproduce the #1723
and the python script will exit before `sys.exit(1)` is called with a
different error.