feat(toolchain): support specifying x.y versions in transitions (#1720)
This is inspired by how rules_go is registering their toolchains.

Their toolchains have multiple `target_settings` values. This
allows for a simpler passing of `X.Y` version to the `py_binary` and
`py_test` rules and does not strictly require us to provide the APIs
that pass the full python version value as the closure. This is only
possible because #1555 introduced working aliases and now we can also
have this.

Summary:
- refactor: move the toolchain_def to starlark as opposed to templating
- refactor: move the version setting as well
- feat: support matching on X.Y versions
- feat: X.Y.Z will match if X.Y is used as python_version flag and the
  MINOR_MAPPING has `"X.Y": "X.Y.Z"`.
- test: add tests checking the generated config settings.
- doc: add an example of how we could use the transition files directly

See
https://github.com/bazelbuild/rules_go/blob/master/go/private/go_toolchain.bzl#L181

---------

Co-authored-by: Richard Levasseur <richardlev@gmail.com>