Ensure regular file bit set in wheel archives (#1731)
#1453 added logic to manually set zipinfo for each file in the wheel
archive. When wheel archives generated by `_Whlfile` are installed by
the `installer` module (or `pip`), the file mode is inspected as such:

* `installer`:
https://github.com/pypa/installer/blob/fcc0d6f14f99974316c2e490cead07a9a0b7a6ac/src/installer/sources.py#L318-L321
* `pip`:
https://github.com/pypa/pip/blob/0f21fb92/src/pip/_internal/utils/unpacking.py#L96-L100

as you can tell, if the regular file bit is not set, the installer
thinks that the file being installed is not an executable and therefore
the executable bit will not be preserved when the file from wheel is
extracted onto the host filesystem.

Since all files being archived into Whlfile are regular files anyway, we
set `S_IFREG` on the file mode for all files in the zip archive.

Fixes #1711

---------

Signed-off-by: Thomas Lam <thomaslam@canva.com>
Co-authored-by: Ignas Anikevicius <240938+aignas@users.noreply.github.com>