feat: Add support for envsubst in extra_pip_args (#1673)
Since `MODULE.bazel` files don't support `load()`, it is more difficult
to grab the value of an environment variable such as `PIP_INDEX_URL` or
`PIP_RETRIES` and feed it to `pip.parse()`. Some of these `PIP_*`
variables are useful in `extra_pip_args`. To enable this use case, also
for the bzlmod scenario, add the `envsubst` attribute, which is a list
of environment variables to substitute in `extra_pip_args`. If
`$VARNAME`, `${VARNAME}` or `${VARNAME:-default}` is found in
`extra_pip_args`, substitute the value of the variable.