docs(gazelle): Add details about python_root directive.
While figuring out how to use gazelle with a `src` dir (see
https://github.com/bazelbuild/rules_python/issues/1775) (I figured it
out by the way, hence this PR), I noticed that the docs were a little
unclear.

This PR updates the `gazelle` docs to include details about the
`python_root` directive and how to use it.

I also fix a the default value for the
`python_generation_mode_per_file_include_init` directive, as it was
incorrectly listed as "package".