fix(internal): add WORKSPACE.bzlmod to ensure dependency separation
Before this PR, WORKSPACE file would be also used when using
'rules_python' under bzlmod and when developing locally it would yield
to inconsistent behaviour between the examples and the main workspace.

This PR ensures that the `dev_pip` is used via `bzlmod`. Whilst there
may be a little bit of duplication from now on, it ensures that
non-bzlmod code paths do not leak into the bzlmod dev setup.