feat: Make Sphinx integration experimental publicly visible
With this commit the visibility settings are adjusted to use the Sphinx
Integration from rules_python also in other (dependent) projects.

Since the Sphinx integration is not yet stable, we denote in the
documentation that this is done under the experimental API support.
Meaning, breaking changes can occure at any point in time.

See the discussion in #1796