feat: Allow additional tools for sphinx_docs
When using some sphinx plugins they require that other (not necessarily
Python) tools are available.

One example is the plugin https://github.com/basejumpa/sphinxcontrib-umlet
which requires that UMLet is somehow within the sandbox.

My understanding is that the right way todo so is the following:

1st: Provide a way how additional tools can be passed to the action
-> Done in this commit via the exposed tools attribute

2nd: Allow a way to configure the path to the tool
-> Done with this commit via the ctx.expand_location for additional
attributes

Both changes should be backward compatible and enable a wider usage of
the Sphinx rules.