refactor: toolchainize py_proto_library (#1577)
Enables use of `--incompatible_enable_proto_toolchain_resolution` flag
that launched in Bazel 7. This allows users to choose a pre-built
`protoc` or use the runtime from https://pypi.org/project/protobuf/
rather than be forced to use hard-coded values in Bazel core.

This change is also happening in other language rulesets that provide
first-class protobuf support, e.g.
https://github.com/bazelbuild/rules_go/issues/3895

No update to CHANGELOG.md in this PR as the feature is not yet
documented for end-users, this just makes it possible to enable the
flag. A follow-up PR will provide user instructions.

---------

Co-authored-by: Alex Eagle <alex@aspect.dev>