refactor: Load cc symbols from @rules_cc
Work towards incompatible_stop_exporting_language_modules, see
https://github.com/bazelbuild/bazel/issues/19455

This elevates rules_cc from a dev dependency to a full dependency of rules_python.

We need to add rules_cc to the internal deps as it has to be loaded before
stardoc. Otherwise, a too old rules_cc version is used when using the legacy
WORKSPACE setup