fix(whl_library): stabilize dependency per python version handling
We start using the recently introduced `is_python_config_setting` to make
it possible to have a working select statement when multiple python
version selection needs to happen in a `whl_library`.

This adds further fixes so that the correct dependencies are pulled in when the
`python_version` string flag is unset thus making this implementation suitable
for `bzlmod` use case where we would use a single `whl_library` instance for
multiple python versions within the hub.

Work towards #735.