feat(pip): support specifying requirements files per (os, arch)
This change implements the necessary selection of the requirement files
based on the (os, arch) for both - bzlmod and legacy workspace
code paths. As part of this addition I have moved some tests from
integration tests to unit tests as now all of the logic on selecting the
right requirement file, ensuring that there are no duplicate requirement
lines and that the user knows what to provide is done in a single place.

This should be a non-breaking change for most users unless they have
been passing `requirements_linux` together with `extra_pip_args =
["--platform=manylinux_2_4_x86_64"]`, in which case they would have to
change their code to use `requirements_lock` attribute which itself is a
trivial change.