feat(toolchains): Add riscv64 linux platform to available toolchain platforms (#1907)
This change adds support for the riscv64 linux platform to rules_python.
It is one of two changes necessary to allow riscv64 linux platforms to
bootstrap modern versions of bazel.

Before this change an attempt to bootstrap bazel would fail with this
error:

```
ERROR: An error occurred during the fetch of repository 'python':
   Traceback (most recent call last):
        File "/home/ubuntu/.cache/bazel/_bazel_ubuntu/c892bce6f2ccb80708b13f24cc8fb67e/external/rules_python/python/private/toolchains_repo.bzl", line 139, column 38, in _toolchain_aliases_impl
                host_platform = get_host_platform(os_name, arch)
        File "/home/ubuntu/.cache/bazel/_bazel_ubuntu/c892bce6f2ccb80708b13f24cc8fb67e/external/rules_python/python/private/toolchains_repo.bzl", line 298, column 13, in get_host_platform
                fail("No platform declared for host OS {} on arch {}".format(os_name, arch))
Error in fail: No platform declared for host OS linux on arch riscv64
```

This change fixes that error.

---------

Co-authored-by: Ignas Anikevicius <240938+aignas@users.noreply.github.com>