feat: add whl_filegroup (#1904)
This adds a `whl_filegroup` rule which can take a wheel as input and
extract it, making the
output available to consumers. A pattern is allowed to better control
what files
are extracted.

This is handy to, e.g. get the numpy headers from the numpy wheel so
they can be
used in a cc_library. e.g., 

Closes #1903 
Fixes #1311