rewrite the filtering algorithm and harden the code
With this we should not get failures, but we would get different
wheels, which from the user point of view might be better experience.

I have also rewrote the tests to better test things and be more succinct.

The complex thing here is that we want to do some filtering of the whls
but we should not do too much a we are doing what the select should be
doing.