fix(py_binary): drop the unused argument.
It seems that there is an extra argument that is there but I am not sure about
the reason, it is most likely a leftover. As part of this change we add an
analysis test for non-windows zip building. I could reproduce the failure and
now the newly added test passes.

Fixes #1954