feat: add runtime_env toolchain suite to replace "autodetecting" toolchain
This adds a more comprehensive replacement for the "autodetecting"
toolchain. Specifically, it defines all our toolchain types so that
they take precedence when specified. This prevents the hemertic
toolchains (registered by default) from accidentally being used when
undesired.

To keep the behaviour backwards compatible, an alias is added for the
autodetecting toolchain with a deprecation notice.

Co-Authored-by: Richard Levasseur <richardlev@gmail.com>