fix(precompile): autodetecting toolchain and feature flags
With this we can finally be correct with how we deal with the precompile
toolchains. This is done in two ways:
1. Add a config_setting that explicitly enableds/disables the precompile
   toolchains.
2. Add an autodetecting toolchain for precompile in cases where users
   may want to use the autodetecting toolchain and we would like to
   ensure that the hermetic toolchain is not pulled in because it exists
   and is used in the target.

To keep the behaviour backwards compatible, add an alias to the
autodetecting toolchain with a deprecation notice. The API is new in
0.33.0 and it is unfortunate that we have to do this, but the users
should still be able to use the code as is.

Fixes #1967.