fix: allow creating PyRuntimeInfo without specifying interpreter_version_info (#1992)
This regression was flagged by some of the Bazel tests. When the
dict->struct interpreter_version code was factored out, it didn't
properly handle the value `None`, which occurs if the provider is
directly instantiated (None comes from the initializer default).

To fix, treat None as an empty dict. The logic already assumes the dict
is potentially empty or missing keys, so it can continue on OK.

Also adds a test for creating PyRuntimeInfo without this argument.