fix: make first default output the executable again (#2010)
This fixes a small change in behavior identified by some Google
regression tests. When precompiling was introduced, the target's
executable was no longer the first file in the default outputs depset.
While that behavior isn't a strong contract, it is the convention with
many other rules, and the existing behavior for Bazel 7+.

To fix, put the executable as the first value in the default outputs
list. Also adds a test for this behavior.