refactor: call a function to define internal pypi flags instead of listcomp
This is to so tooling can more automatically process the files. In
particular, it helps tools, like buildozer process the files, which
makes it easier to import the code into Google. This is because there is
a named target that buildozer can be told to process, whereas, with a
list comprehension, it's an arbitrary chunk of code that has to be
patched, without an identifiable label.