feat: add runtime_env toolchain suite to replace "autodetecting" toolchain (#2018)
This adds a more comprehensive replacement for the "autodetecting"
toolchain. Specifically, it defines all our toolchain types so that they
take precedence when specified. This prevents the hermetic toolchains
(registered by default) from accidentally being used when undesired.

To keep the behavior backwards compatible, an alias is added for the
autodetecting toolchain with a deprecation notice.

The name `runtime_env` was chosen instead of "autodetecting" so that
it's more clear these
toolchains are not "automatic" or "detecting" anything -- they're just
taking a value
from the runtime environment and using it.

---------

Co-authored-by: Ignas Anikevicius <240938+aignas@users.noreply.github.com>