fix: use `exec` to invoke the stage-2 bootstrap for non-zip case (#2047)
When the two-stage bootstrap is used, the parent shell process runs
python as a child
process, which changes how signals are propagated. Specifically, if a
signal is sent
_directly_ to the parent (e.g. `kill $parent`), the child process
(python) won't receive
it and it will appear to be ignored. This is because the parent process
is busy waiting
for the child process.

To fix, invoke the python process using `exec` instead. Because the
process is entirely
replaced, signals are sent directly to the replacement. This can't be
used for zip files,
though, because they rely on a catching the exit signal to perform
cleanup of the extracted
files.

Fixes https://github.com/bazelbuild/rules_python/issues/2043

---------

Co-authored-by: Richard Levasseur <rlevasseur@google.com>