fix: make mac zip builds work (#2052)
Macs have an older version of `mktemp`, one that doesn't support the
`--suffix` arg. This
caused the combination of Macs and `--build_python_zip
--bootstrap_impl=script` to fail.

To fix, remove the `--suffix` arg. As far as I can tell, the suffix
string,
"Bazel.runfiles_", is just informational, so this is fine to remove.

Also adds tests to verify that a binary runs with/without zip and for
the script
bootstrap.

Fixes https://github.com/bazelbuild/rules_python/issues/2030