fix: watch_srcs and more things in repo_utils.
Move more things to repo_utils and stop using watch_tree, because it
also starts watching __pycache__ files, that is causing excesive
restarts. What is more, we previously where using uname to get the OS,
but it does not make sense to do that, we can use rctx.os.name instead
and common utilities.

What is more, be more explicit about what OSes are supported in the
local_runtime_repo.